+++ src/dotnet/APIView/APIViewWeb/DTOs/SiteNotificationDto.cs
@@ -1,3 +1,4 @@
+using System;
 using APIViewWeb.Models;
 
 namespace APIViewWeb.DTOs;
@@ -9,7 +10,29 @@ public class SiteNotificationDto
     public string Title { get; set; }
     public string Summary { get; set; }
     public string Message { get; set; }
-    public string Status { get; set; }
+
+    private string _status;
+    public string Status
+    {
+        get => _status;
+        set => _status = NormalizeStatus(value);
+    }
+
+    private static string NormalizeStatus(string value)
+    {
+        if (string.IsNullOrWhiteSpace(value))
+            throw new ArgumentException("Status cannot be null or empty.");
+
+        value = value.Trim().ToLowerInvariant();
+        return value switch
+        {
+            SiteNotificationStatus.Info => SiteNotificationStatus.Info,
+            SiteNotificationStatus.Success => SiteNotificationStatus.Success,
+            SiteNotificationStatus.Error => SiteNotificationStatus.Error,
+            SiteNotificationStatus.Warning => SiteNotificationStatus.Warning,
+            _ => throw new ArgumentException($"Invalid status value: {value}")
+        };
+    }
 }
 
 public static class SiteNotificationStatus
